Add ListenerTest
There is two tests, one that tests that the fixture is set when a new
class is executed. The other tests that the fixture is not set, if it's
another type of test descriptor, like a method or package.

Also had to change from implementation to testImplementation
so that intTest have access to assertj.